refactor: add viewmixin, openaipi module, change flake8 rules and db models
