+++ content/api/views.py
@@ -19,6 +19,7 @@ from rest_framework import mixins, status, viewsets
 from rest_framework.response import Response
 
 from content import filters
+from content.mixins import MultiSerializerViewSetMixin
 from content.models import (
     AboutUsVideo,
     Chapter,
@@ -140,7 +141,9 @@ class AboutUsVideoViewSet(viewsets.GenericViewSet):
         """,
     ),
 )
-class TargetedFundraisingViewSet(viewsets.ReadOnlyModelViewSet):
+class TargetedFundraisingViewSet(
+    MultiSerializerViewSetMixin, viewsets.ReadOnlyModelViewSet
+):
     """Информация об адресных сборах.
 
     Используйте этот эндпоинт, чтобы отобразить информацию об адресных сборах.
@@ -148,16 +151,10 @@ class TargetedFundraisingViewSet(viewsets.ReadOnlyModelViewSet):
     """
 
     queryset = TargetedFundraising.objects.all()
-
-    def get_serializer_class(self):
-        """Выбирает сериализатор в зависимости от действия.
-
-        Возвращает краткий сериализатор для списка (list)
-        и детальный для отдельного сбора (retrieve).
-        """
-        if self.action == 'retrieve':
-            return serializers.TargetedFundraisingDetailSerializer
-        return serializers.TargetedFundraisingListSerializer
+    serializer_classes = {
+        'list': serializers.TargetedFundraisingListSerializer,
+        'retrieve': serializers.TargetedFundraisingDetailSerializer,
+    }
 
 
 @extend_schema(tags=['Employees group'])
@@ -169,7 +166,9 @@ class TargetedFundraisingViewSet(viewsets.ReadOnlyModelViewSet):
         summary='Получить карточку Сотрудника по ID.',
     ),
 )
-class EmployeeViewSet(viewsets.ReadOnlyModelViewSet):
+class EmployeeViewSet(
+    MultiSerializerViewSetMixin, viewsets.ReadOnlyModelViewSet
+):
     """Информация о сотрудниках.
 
     Используйте этот эндпоинт, чтобы отобразить информацию о сотрудниках.
@@ -177,16 +176,10 @@ class EmployeeViewSet(viewsets.ReadOnlyModelViewSet):
     """
 
     queryset = Employee.objects.all()
-
-    def get_serializer_class(self):
-        """Выбирает сериализатор в зависимости от действия.
-
-        Возвращает краткий сериализатор для списка (list)
-        и детальный для отдельного сотрудника (retrieve).
-        """
-        if self.action == 'retrieve':
-            return serializers.EmployeeDetailSerializer
-        return serializers.EmployeeSerializer
+    serializer_classes = {
+        'list': serializers.EmployeeSerializer,
+        'retrieve': serializers.EmployeeDetailSerializer,
+    }
 
 
 @extend_schema(tags=['Projects group'])
@@ -244,7 +237,7 @@ class MissionViewSet(
         summary='Получить Новость по ID.',
     ),
 )
-class NewsViewSet(viewsets.ReadOnlyModelViewSet):
+class NewsViewSet(MultiSerializerViewSetMixin, viewsets.ReadOnlyModelViewSet):
     """Получить список Новостей, или конкретную по её ID."""
 
     queryset = News.objects.select_related('project').prefetch_related(
@@ -252,16 +245,10 @@ class NewsViewSet(viewsets.ReadOnlyModelViewSet):
     )
     filter_backends = [DjangoFilterBackend]
     filterset_class = filters.NewsFilter
-
-    def get_serializer_class(self):
-        """Выбирает сериализатор в зависимости от действия.
-
-        Возвращает краткий сериализатор для списка (list)
-        и детальный для отдельного сотрудника (retrieve).
-        """
-        if self.action == 'retrieve':
-            return serializers.NewsDetailSerializer
-        return serializers.NewsSerializer
+    serializer_classes = {
+        'list': serializers.NewsSerializer,
+        'retrieve': serializers.NewsDetailSerializer,
+    }
 
 
 @extend_schema(tags=['Directions group'])
@@ -329,17 +316,13 @@ class CoachingViewSet(viewsets.ReadOnlyModelViewSet):
         summary='Получить Вакансию по ID.',
     ),
 )
-class VacancyViewSet(viewsets.ReadOnlyModelViewSet):
+class VacancyViewSet(
+    MultiSerializerViewSetMixin, viewsets.ReadOnlyModelViewSet
+):
     """Получить список Вакансий, или конкретную по её ID."""
 
     queryset = Vacancy.objects.all()
-
-    def get_serializer_class(self):
-        """Выбирает сериализатор в зависимости от действия.
-
-        Возвращает краткий сериализатор для списка (list)
-        и детальный для отдельного сотрудника (retrieve).
-        """
-        if self.action == 'retrieve':
-            return serializers.VacancyDetailSerializer
-        return serializers.VacancySerializer
+    serializer_classes = {
+        'list': serializers.VacancySerializer,
+        'retrieve': serializers.VacancyDetailSerializer,
+    }
