+++ content/models/about_us_video.py
@@ -31,7 +31,7 @@ class AboutUsVideo(TimestampMixin, models.Model):
 
     def __str__(self):
         """Возвращает строковое представление объекта видео."""
-        return self.title if self.title is not None else EMPTY_VALUE_DISPLAY
+        return self.title or EMPTY_VALUE_DISPLAY
 
     @classmethod
     def get_solo(cls):
