+++ form_sender/views.py
@@ -11,9 +11,6 @@ from django.core.mail import send_mail
 
 from drf_spectacular.utils import (
     extend_schema,
-    extend_schema_view,
-    OpenApiExample,
-    OpenApiResponse,
 )
 
 from rest_framework import status
@@ -21,75 +18,14 @@ from rest_framework.response import Response
 from rest_framework.throttling import ScopedRateThrottle
 from rest_framework.views import APIView
 
+from .openapi import feedback_form_schema
 from .serializers import (
     FeedbackFormSerializer,
-    FeedbackSuccessResponseSerializer,
-    FeedbackErrorResponseSerializer,
-    ThrottleErrorResponseSerializer,
 )
 
 
 @extend_schema(tags=['Forms group'])
-@extend_schema_view(
-    post=extend_schema(
-        summary='Отправить форму обратной связи',
-        description=(
-            'Отправляет форму обратной связи и уведомляет администратора по '
-            'email.\n\n'
-            '**Лимит запросов:**\n'
-            '• 100 запросов в день\n'
-        ),
-        request=FeedbackFormSerializer,
-        responses={
-            200: OpenApiResponse(
-                response=FeedbackSuccessResponseSerializer,
-                description='Форма успешно обработана и сообщение отправлено',
-                examples=[
-                    OpenApiExample(
-                        'Успешная отправка',
-                        value={
-                            'status': 'success',
-                            'message': 'Сообщение отправлено успешно!',
-                        },
-                    )
-                ],
-            ),
-            400: OpenApiResponse(
-                response=FeedbackErrorResponseSerializer,
-                description='Ошибки валидации входных данных',
-                examples=[
-                    OpenApiExample(
-                        'Ошибки валидации',
-                        value={
-                            'name': ['Обязательное поле.'],
-                            'phone_number': [
-                                'Неверный формат номера телефона'
-                            ],
-                            'message': ['Обязательное поле.'],
-                            'accept_terms': [
-                                'Необходимо принять условия обработки '
-                                'персональных данных'
-                            ],
-                        },
-                    )
-                ],
-            ),
-            429: OpenApiResponse(
-                response=ThrottleErrorResponseSerializer,
-                description='Превышен лимит запросов (rate limiting)',
-                examples=[
-                    OpenApiExample(
-                        'Rate limit exceeded',
-                        value={
-                            'detail': 'Request was throttled. Expected '
-                            'available in 86400 seconds.'
-                        },
-                    )
-                ],
-            ),
-        },
-    )
-)
+@feedback_form_schema
 class FeedbackFormView(APIView):
     """Форма обратной связи."""
